+++ BoardingPassExample/Onboarding Demo/Process View Controllers/ActionableViewController.swift
@@ -9,7 +9,7 @@
 import UIKit
 import BoardingPass
 
-@objc protocol Actions: NSObjectProtocol {
+@objc protocol NavigationActionResponder: NSObjectProtocol {
     func handleNextTapped(_ sender: UIResponder)
     func handleSkipTapped(_ sender: UIResponder)
 }
@@ -20,9 +20,9 @@ class ActionableViewController: UIViewController {
         navigationItem.backBarButtonItem = UIBarButtonItem.backButton
         navigationItem.rightBarButtonItem = UIBarButtonItem.skipButton
         navigationItem.rightBarButtonItem?.target = nil
-        navigationItem.rightBarButtonItem?.action = #selector(Actions.handleSkipTapped)
+        navigationItem.rightBarButtonItem?.action = #selector(NavigationActionResponder.handleSkipTapped)
         let nextButton = UIButton(title: NSLocalizedString("Next", comment: "Next button title"), font: .onboardingFont)
-        nextButton.addTarget(nil, action: #selector(Actions.handleNextTapped), for: .touchUpInside)
+        nextButton.addTarget(nil, action: #selector(NavigationActionResponder.handleNextTapped), for: .touchUpInside)
         view.addSubview(nextButton)
         let constraints: [NSLayoutConstraint] = [
             nextButton.centerXAnchor.constraint(equalTo: view.centerXAnchor),
