+++ BoardingPassExample/Onboarding Demo/Process View Controllers/FirstViewController.swift
@@ -21,40 +21,3 @@ class FirstViewController: ActionableViewController {
         animate(color: .darkGray)
     }
 }
-
-extension Progress {
-    convenience init(completedUnitCount: Int64, totalUnitCount: Int64) {
-        self.init(totalUnitCount: totalUnitCount)
-        self.completedUnitCount = completedUnitCount
-    }
-}
-
-extension UIButton {
-
-    convenience init(title: String, font: UIFont) {
-        self.init(type:.system)
-        setTitle(title, for: UIControlState())
-        titleLabel?.font = font
-    }
-
-}
-
-extension UIBarButtonItem {
-    static var backButton: UIBarButtonItem {
-         return UIBarButtonItem(title: NSLocalizedString("Back", comment: "generic back button title"), style: .plain, target: nil, action: nil)
-    }
-
-    static var skipButton: UIBarButtonItem {
-        return UIBarButtonItem(title: NSLocalizedString("Skip", comment: "generic skip button title"), style: .plain, target: nil, action: nil)
-    }
-}
-
-extension UIViewController {
-    func animate(color: UIColor) {
-        perform(coordinatedAnimations: { (vc, _) -> (() -> Void) in
-            return {
-                vc?.view.backgroundColor = color
-            }
-        })
-    }
-}
